:recycle: Hide infer_layer
