+++ tests/test_pflow.py
@@ -2,7 +2,7 @@ from __future__ import annotations
 
 import networkx as nx
 import pytest
-from swiflow import common, pflow
+from swiflow import _common, pflow
 from swiflow.common import PPlane
 
 from tests.assets import CASES, FlowTestCase
@@ -41,5 +41,5 @@ def test_infer_verify(c: FlowTestCase) -> None:
     if c.pflow is None:
         pytest.skip()
     f, _ = c.pflow
-    layer = common.infer_layer(c.g, f, pplane=c.pplane)
+    layer = _common.infer_layer(c.g, f, pplane=c.pplane)
     pflow.verify((f, layer), c.g, c.iset, c.oset, pplane=c.pplane)
